requirements.txt
streamlit==1.28.0
torch
torchvision
pillow
numpy